test: fix tests that were relying on a timeout instead of an immediate exception
(cherry picked from commit 241bc0b853f1427103ce73269a4b58b61bae5d1f)
